Character animated in all directions
