+++ covid_game/public/index.html
@@ -29,7 +29,7 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-        <center><canvas id="gameScreen" width="300" height="200"></canvas></center>
+        <center><canvas id="gameScreen" width="1500" height="800"></canvas></center>
         <img id="sprite" src="./assets/sprite.png"/>
 
   </body>
